Added pr review and lint formatting
