+++ cogs/clash_river_war.py
@@ -2,11 +2,10 @@ import discord
 from discord.ext import commands
 from textwrap import dedent
 from common import schemas
-from common import schemas
-from textwrap import dedent
 import logging
 from royale_api_website_scraper import RoyaleApiWebsiteScraper
 from clan_members_rank import ClanMembersRanker
+
 
 def create_clan_members_ranked_embed(clan_info, clan_members_ranked, card_type_arg='all'):
     n = 20  # Number of players to show
@@ -38,6 +37,7 @@ def create_clan_members_ranked_embed(clan_info, clan_members_ranked, card_type_a
 
     return embed
 
+
 def boat_attackers_embed(boat_attackers, clan_info):
     if len(boat_attackers) == 0:
         description = 'There are no clan members who attacked enemy boats this week.'
@@ -61,7 +61,8 @@ def boat_attackers_embed(boat_attackers, clan_info):
 
     return embed
 
-class Admin(commands.Cog):
+
+class ClashRiverWar(commands.Cog):
     def __init__(self, client):
         self.client = client
         self.clan_members_ranker = ClanMembersRanker()
@@ -87,30 +88,6 @@ class Admin(commands.Cog):
             return False
         else:
             return True
-    @commands.command(name="bothelp", pass_context=True)
-    async def bothelp(self, ctx):
-        async with ctx.typing():
-            embed = discord.Embed(
-                description=dedent('''
-                This bot provides a variety of useful commands to help with clan management.
-                '''),
-                colour=discord.Colour.purple()
-            )
-            url = 'https://royaleapi.com/clan/9GULPJ9L/war/race'
-            ausclan_cmds = [
-                f'[!ausclan [filter]]({url}) - Shows clan members ranked by number of high level cards they have. [filter] '
-                'can be one of \'all\', \'troop\', \'spell\' or \'building\'. Defaults to \'all\' if not specified',
-                f'[!ausclanboat]({url}) - Lists clan members who attacked enemy boats this week.',
-            ]
-            generic_cmds = [
-                f'[!bothelp]({url}) - Shows this help message about available commands.',
-                f'[!membercardsranked [clan_tag] [filter]]({url}) - Does same thing as `!ausclan` but '
-                'for the specific [clan_tag] instead.',
-                f'[!boatattack [clan_tag]]({url}) - Does same thing as `!ausclanboat` but for the specific [clan_tag] instead.'
-            ]
-            embed.add_field(name='Ausclan commands', value='\n'.join(ausclan_cmds), inline=False)
-            embed.add_field(name='Other commands', value='\n'.join(generic_cmds), inline=False)
-        await ctx.send(embed=embed)
 
     @commands.command(name="membercardsranked", pass_context=True)
     async def membercardsranked(self, ctx, *args):
@@ -150,5 +127,6 @@ class Admin(commands.Cog):
     async def ausclanboat(self, ctx):
         await self.boatattack(ctx, '9GULPJ9L')
 
+
 def setup(client):
-    client.add_cog(Admin(client))
+    client.add_cog(ClashRiverWar(client))
