+++ tests/resources/clash_royale_client_responses.py
